\name{element_clear}
\alias{element_clear}
\title{Send keys}
\usage{
  element_clear(ID = NULL)

  element_clear(ID = NULL)
}
\arguments{
  \item{ID}{an ID element}

  \item{ID}{an ID element}
}
\description{
  Send a sequence of key strokes to the active element

  Clear a textarea or text input element's value.
}

