\name{element_click}
\alias{element_click}
\title{Click element}
\usage{
  element_click(ID = ID, times = 1, button = "left")
}
\arguments{
  \item{ID}{an ID element}

  \item{times}{numerical value indicating the number of
  clicks}

  \item{button}{a string indicating left or right mouse
  button}
}
\description{
  Click on a specified element
}

