\name{element_css_find}
\alias{element_css_find}
\title{Find an element via CSS selector}
\usage{
  element_css_find(value)
}
\arguments{
  \item{value}{a string specifying a CSS selector}
}
\value{
  ID element(s) that match the CSS selector
}
\description{
  Search for an element on the page via XPath, starting
  from the document root
}

