\name{element_ptext_find}
\alias{element_ptext_find}
\title{Find an element via partial text matching}
\usage{
  element_ptext_find(value)
}
\arguments{
  \item{value}{a string of text that is searched for in the
  page}
}
\value{
  ID element(s) whose visible text partially matches the
  search value
}
\description{
  Search for an element on the page via CSS selector,
  starting from the document root
}

