\name{element_xpath_find}
\alias{element_xpath_find}
\title{Find an element via XPath}
\usage{
  element_xpath_find(value)
}
\arguments{
  \item{value}{a string specifying a XML Path Language
  statement}
}
\value{
  ID element(s) that match the XPath expression
}
\description{
  Search for an element on the page, starting from the
  document root
}

