\name{get.window_position}
\alias{get.window_position}
\title{Get current position of a window}
\usage{
  get.window_position(handle = NULL)
}
\arguments{
  \item{handle}{a string indicating the handle of the
  window to be maximized}
}
\value{
  a vector (of length 2) specifying x and y coordinates of
  the window
}
\description{
  Get the position of the specified window
}

