\name{implicit_wait}
\alias{implicit_wait}
\title{Set implicit waiting time}
\usage{
  implicit_wait(ms = 0)
}
\arguments{
  \item{ms}{numerical value indicating the amount of time
  (in milliseconds) to wait for elements to appear on the
  page}
}
\description{
  Set the amount of time the driver should wait when
  searching for elements
}

