\name{liv}
\alias{liv}
\title{Locate and determine element's screen position}
\usage{
  liv(ID = ID)
}
\arguments{
  \item{ID}{an ID element on which the click is performed}
}
\value{
  a vector (of length 2) specifying x and y coordinates of
  the window
}
\description{
  Determine an element's location on the screen once it has
  been scrolled into view
}

