\name{post.window_position}
\alias{post.window_position}
\title{Change position of a window}
\usage{
  post.window_position(position = NULL, handle = NULL)
}
\arguments{
  \item{position}{a numeric vector (of length 2) that
  specifies x and y coordinates for the new position}

  \item{handle}{a string indicating the handle of the
  window to be maximized}
}
\description{
  Get the size of the specified window.
}

