\name{post.window_size}
\alias{post.window_size}
\title{Resize window}
\usage{
  post.window_size(size = NULL, handle = NULL)
}
\arguments{
  \item{size}{a numeric vector (of length 2) that specifies
  width and height of the window}

  \item{handle}{a string indicating the handle of the
  window to be maximized}
}
\description{
  Change the size of the specified window
}

