\name{window_change}
\alias{window_change}
\title{Change window focus}
\usage{
  window_change(handle = NULL)
}
\arguments{
  \item{handle}{a string indicating the handle of the
  window to be maximized}
}
\description{
  Change focus to another window. The window to change
  focus to may be specified by its server assigned window
  handle, or by the value of its name attribute
}

